Don't compile cgroups stuff on not-linux
